Change name utils.py to signals_bot.py, refactored imports
